Create a base definition for HPCG
This commit creates a base application definition for HPCG, to allow
NVIDIA HPCG to inherit from this as well.
